refactor: Rename inference code
